Completed basic level handling
Collision Detection Script works as intended to reload current levels if colliding with non-landing pad objects and loading to the next level if colliding with landing pad objects.

The collision detection script also loads the first scene in the build index (Level 1) if the next scene would be out of range of the current number of scenes in the build index.