Modified Prefabs
Theory: Synchronization of project was prevented because of discrepancy between prefab objects and the existence of their prefabs in the project.

Unpacked any non-essential/indirect assets that were under another prefab or that were a part of level design so that the project might properly synchronize when working with the other device.